+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Operator::apply` method to try to construct an AST (convenient alternative to `try_into`/`try_from`).
 
 ## [0.5.0] - 2020-11-29
 ### Added